Moved incoming-message to its own file
